chore(core): cleanup + pr feedback
