+++ packages/nx/src/project-graph/plugins/plugin-pool.ts
@@ -1,8 +1,8 @@
 import { ChildProcess, fork } from 'child_process';
 import path = require('path');
-import { PluginWorkerResult, consumeMessage, createMessage } from './types';
+import { PluginWorkerResult, consumeMessage, createMessage } from './messaging';
 import { PluginConfiguration } from '../../config/nx-json';
-import { RemotePlugin, nxPluginCache } from './nx-plugin';
+import { RemotePlugin, nxPluginCache } from './internal-api';
 import { ProjectGraph } from '../../config/project-graph';
 import { logger } from '../../utils/logger';
 
@@ -11,26 +11,33 @@ const pool: ChildProcess[] = [];
 const pidMap = new Map<number, string>();
 
 export function loadRemoteNxPlugin(plugin: PluginConfiguration, root: string) {
-  const isTest = path.extname(__filename) === '.ts';
+  // this should only really be true when running unit tests within
+  // the Nx repo. We still need to start the worker in this case,
+  // but its typescript.
+  const isWorkerTypescript = path.extname(__filename) === '.ts';
   const workerPath = path.join(__dirname, 'plugin-worker');
   const worker = fork(workerPath, [], {
     stdio: ['ignore', 'inherit', 'inherit', 'ipc'],
     env: {
       ...process.env,
-      ...(isTest
+      ...(isWorkerTypescript
         ? {
+            // Ensures that the worker uses the same tsconfig as the main process
             TS_NODE_PROJECT: path.join(__dirname, '../../../tsconfig.lib.json'),
           }
         : {}),
     },
     execArgv: [
       ...process.execArgv,
-      ...(isTest ? ['-r', 'ts-node/register'] : []),
+      // If the worker is typescript, we need to register ts-node
+      ...(isWorkerTypescript ? ['-r', 'ts-node/register'] : []),
     ],
   });
   worker.send(createMessage({ type: 'load', payload: { plugin, root } }));
   pool.push(worker);
+  
   logger.verbose(`[plugin-worker] started worker: ${worker.pid}`);
+  
   return new Promise<RemotePlugin>((res, rej) => {
     worker.on('message', createWorkerHandler(worker, res, rej));
     worker.on('exit', () => workerOnExitHandler(worker));
@@ -40,17 +47,23 @@ export function loadRemoteNxPlugin(plugin: PluginConfiguration, root: string) {
 let pluginWorkersShutdown = false;
 
 export async function shutdownPluginWorkers() {
+  // Clears the plugin cache so no refs to the workers are held
   nxPluginCache.clear();
+
+  // Marks the workers as shutdown so that we don't report unexpected exits
   pluginWorkersShutdown = true;
+  
   const promises = [];
+
   for (const p of pool) {
     p.send(createMessage({ type: 'shutdown', payload: undefined }), (error) => {
       if (error) {
         // This occurs when the worker is already dead, and we can ignore it
       } else {
         promises.push(
+          // Create a promise that resolves when the worker exits
           new Promise<void>((res, rej) => {
-            p.on('exit', () => res());
+            p.once('exit', () => res());
           })
         );
       }
@@ -59,11 +72,28 @@ export async function shutdownPluginWorkers() {
   return Promise.all(promises);
 }
 
+
+/**
+ * Creates a message handler for the given worker.
+ * @param worker Instance of plugin-worker
+ * @param onload Resolver for RemotePlugin promise
+ * @param onloadError Rejecter for RemotePlugin promise
+ * @returns Function to handle messages from the worker
+ */
 function createWorkerHandler(
   worker: ChildProcess,
   onload: (plugin: RemotePlugin) => void,
   onloadError: (err?: unknown) => void
 ) {
+
+  // We store resolver and rejecter functions in the outer scope so that we can
+  // resolve/reject the promise from the message handler. The flow is something like:
+  // 1. plugin api called
+  // 2. remote plugin sends message to worker, creates promise and stores resolver/rejecter
+  // 3. worker performs API request
+  // 4. worker sends result back to main process
+  // 5. main process resolves/rejects promise based on result
+
   let createNodesResolver: (
     result: Awaited<ReturnType<RemotePlugin['createNodes'][1]>>
   ) => void | undefined;
